+++ src/app/shared/error-modal/error-modal.ts
@@ -1,8 +1,10 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 
+import { TranslatePipe } from '../../core/i18n/translate.pipe';
+
 @Component({
   selector: 'app-error-modal',
-  imports: [],
+  imports: [TranslatePipe],
   templateUrl: './error-modal.html',
   styleUrl: './error-modal.scss',
 })